Bug 26805: Remove remaining instances of jquery.checkboxes plugin from the OPAC
This patch removes the use of the jquery.checkboxes plugin from the 5
templates in the OPAC which still use it. The plugin itself is removed.

To test, apply the patch and test the "Select all" and "Clear all" links
work to select and deselect checkboxes on the following pages:

- The catalog search results page: The controls should also enable
  and disable the "Select titles to..." menu.
- The list contents page:  The controls should also enable
  and disable the "Select titles to..." links.
- The suggestions page: The controls should also enable
  and disable the "Select suggestions to..." link.
- The search history page: The controls should affect only the
  corresponding table ("Current" or "Previous") and should enable and
  disable the "Select searches to..." link.

This patch also modifies the template for grouped search results, as
enabled by the OpacGroupResults system preference, but I don't know how
to test it.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>