Bug 26806: Remove the jquery.checkboxes plugin from the staff client
This patch removes the jquery.checkboxes plugin from the staff client.

It also removes some leftover instances of the "checkboxed" class which
was used in some cases to enable the checkbox plugin on selected
elements.

To test, apply the patch and check that the jquery.checkboxes plugin no
longer exists in /intranet-tmpl/lib/jquery/plugins.

To test changes to the templates, confirm that "Check all" and "Check
none" links still work on these pages:

- The tags review page. See the test plan for Bug 26120.
- The inventory page. See the test plan for Bug 26150.
- The list contents page. See the test plan for Bug 26204.

The other modified templates contained the "checkboxed" class but it
served no purpose:

- Serials -> Serials collection
- Labels -> Edit label batch
- Patron cards -> Edit card batch

If you search the Koha codebase for instances of "checkboxed" the only
results should be from the OPAC (if Bug 26805 hasn't been applied) and
in Z39.50 templates where the class is still used by JavaScript.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>