Bug 26816: Chomp whitespace around i18n strings
This patch adds the chomp shorthand [%- -%] to the output of i18n.inc to
prevent extra spaces from being inserted.

To test, apply the patch and perform a catalog search in the staff
interface which will return results with multiple items.

In the items column, the availability information should not have
unwanted spaces:

3 items, 3 available:

...instead of

3 items , 3 available:

Signed-off-by: Barbara Johnson <barbara.johnson@bedfordtx.gov>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>