Bug 26825: Add span for publication date in OPAC
This patch modifies the XSL templates for OPAC results and OPAC details
so that the parts of the publisher details are wrapped separate <span>s:
Publisher place, name, and date.

To test, apply the patch "restart_all" if necessary.

You can add this to the OPACUserCSS system preference to better see the
change:

.publisher_place { color: red; }
.publisher_name { color: green; }
.publisher_date { color: orange; }

- Peform a search in the OPAC
- In the search results, confirm that the publisher plac, name, and date
  are wrapped in separate <span>s, either by confirming the custom
  colors or by inspecting the source.
- View the details for one of the results. Confirm that the same change
  has been made to the detail page.

Signed-off-by: B Johnson <barbara.johnson@bedfordtx.gov>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>