Bug 26889: Remove extra space from "Damaged :" in item search
This patch adds the chomp shorthand [%- -%] to the BLOCK for showing
labels in the item search template. This collapses the whitespace around
the label, eliminating the unwanted space around the ":"

To test, apply the patch and go to the item search page in the staff
interface. There should be no space before the colon in the "Damaged"
label.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>