Bug 26950: Make Anonymous patron warnings 'errors' and refer to the system information tab on about.pl
Currently checkins can fail for one of two known reasons:
1 - Duplicate issue ids
2 - Anonymous borrower configuration issues

In the first case we refer to the 'data problems' or 'data corrupted' section of the about page. All errors are on the 'System information' tab and this is what we should mention

In the second case we refer to the data problems section, but information is in another section on the tab. Additionally, during checkin we issue an 'Error' but on the about page we call it a 'Warning' we should upgrade to error and specify what is broken

Additionally this patch fixes the warning that should display while issuing to a patron if anonymous is not set and patron has selected privacy

To test:
 1 - Enable system preference OPACPrivacy
 2 - Set system preference AnonymousPatron to '0'
 3 - Log in a a patron on OPAC and set your privacy to never
 4 - Issue a book to that patron in the staff client
 5 - Note there is no warning
 6 - Return the book
 7 - Note the circulation page displays an error and the checkin fails and refers to 'data problems'
 8 - Apply patch
 9 - Issue another book to patron
10 - Note warning on the circulation page
11 - Try to check in a book again
12 - Check the new message makes sense
13 - Click the link to 'about.pl' check that errors are displayed
14 - Check the logs
15 - Confirm updated error message makes sense

Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>
Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>