Bug 27125: Show authority type also for UNIMARC in authority search result display
Bug 12533 added authority type in authority search result display.
But it does not work for UNIMARC, due to :
https://git.koha-community.org/Koha-community/Koha/src/commit/be9f64b638335cac7465ca302e187c7914534fbe/C4/AuthoritiesMarc.pm#L798

I propose to use result->{authtype}

Test plan :
1) Use UNIMARC catalogue
2) Search for autorities
3) Check you see autority type in column
4) Same for a MARC21 catalogue

Signed-off-by: Andreas Roussos <a.roussos@dataly.gr>
I followed the test plan, and the patch works as expected for both
UNIMARC and MARC21. There is actually an improvement in the MARC21 case:
previously, the 'Authority type' column for records in the Default
framework would not display anything, whereas now it displays 'Default'.

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>