Bug 26397: [20.05.x] opac.scss calls non-existent image
This patch removes some obsolete CSS from the OPAC which calls a
non-existent image. There should be no visible change.

This patch is specifically for Koha 20.05.x. The bug does not appear in
master.

To test, apply the patch and rebuild the OPAC SCSS
(https://wiki.koha-community.org/wiki/Working_with_SCSS_in_the_OPAC_and_staff_client).

View the OPAC and confirm that nothing has changed in the style of the
Koha logo in the header.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Nick Clemens <nick@bywatersolutions.com>

Signed-off-by: Andrew Fuerste-Henry <andrew@bywatersolutions.com>